Lots of changes

+++ src/components/Item.js
@@ -14,6 +14,7 @@ import { enGB } from "date-fns/esm/locale";
 import {
   archiveItem,
   refileItem,
+  updateItemDescription,
   completeItem,
   deleteItem,
   undeleteItem,
@@ -34,7 +35,7 @@ const ItemContainer = styled.div`
   grid-template-areas:
     "type body body body body body body body body tag"
     "scheduled scheduled scheduled scheduled . . due due due due";
-  height: 50px;
+  height: ${props => (props.itemType == "TODO" ? "50px" : "30px")};
   width: 650px;
   border: 1px solid;
   border-color: ${props => props.theme.colours.borderColour};
@@ -107,7 +108,9 @@ class Item extends Component {
     this.state = {
       projectDropdownVisible: false,
       scheduledDatePopupVisible: false,
-      dueDatePopupVisible: false
+      dueDatePopupVisible: false,
+      descriptionEditable: false,
+      preventDefaultEvents: false
     };
     this.setScheduledDate = this.setScheduledDate.bind(this);
     this.setDueDate = this.setDueDate.bind(this);
@@ -116,6 +119,9 @@ class Item extends Component {
 
     this.hotkeyHandler = {
       SET_SCHEDULED_DATE: event => {
+        if (this.state.preventDefaultEvents == true) {
+          return;
+        }
         this.setState({
           scheduledDatePopupVisible: !this.state.scheduledDatePopupVisible,
           dueDatePopupVisible: false,
@@ -123,6 +129,9 @@ class Item extends Component {
         });
       },
       SET_DUE_DATE: event => {
+        if (this.state.preventDefaultEvents == true) {
+          return;
+        }
         this.setState({
           dueDatePopupVisible: !this.state.dueDatePopupVisible,
           scheduledDatePopupVisible: false,
@@ -130,37 +139,88 @@ class Item extends Component {
         });
       },
       ARCHIVE: event => {
+        if (this.state.preventDefaultEvents == true) {
+          return;
+        }
         this.props.archiveItem(event.target.id);
       },
       COMPLETE: event => {
+        if (this.state.preventDefaultEvents == true) {
+          return;
+        }
         this.props.completeItem(event.target.id);
       },
       UNCOMPLETE: event => {
+        if (this.state.preventDefaultEvents == true) {
+          return;
+        }
         this.props.uncompleteItem(event.target.id);
       },
       DELETE: event => {
+        if (this.state.preventDefaultEvents == true) {
+          return;
+        }
         this.props.deleteItem(event.target.id);
       },
       UNDELETE: event => {
+        if (this.state.preventDefaultEvents == true) {
+          return;
+        }
         this.props.undeleteItem(event.target.id);
       },
       REFILE: event => {
+        if (this.state.preventDefaultEvents == true) {
+          return;
+        }
         this.setState({
           projectDropdownVisible: !this.state.projectDropdownVisible,
           dueDatePopupVisible: false,
           scheduledDatePopupVisible: false
         });
       },
+      ENTER: event => {
+        this.setState(
+          {
+            descriptionEditable: false,
+            preventDefaultEvents: false
+          },
+          () => {
+            this.description.blur();
+          }
+        );
+        // TODO: This is kinda silly because the TEXT expects the type at the front
+        this.props.updateItemDescription(
+          event.target.parentElement.id,
+          this.props.type + " " + event.target.innerText
+        );
+      },
+      EDIT: event => {
+        if (this.state.preventDefaultEvents == true) {
+          return;
+        }
+        this.setState(
+          {
+            descriptionEditable: true,
+            preventDefaultEvents: true
+          },
+          () => {
+            this.description.focus();
+          }
+        );
+        event.preventDefault();
+      },
       ESCAPE: event => {
+        this.description.blur();
         this.setState({
           projectDropdownVisible: false,
           dueDatePopupVisible: false,
-          scheduledDatePopupVisible: false
+          scheduledDatePopupVisible: false,
+          descriptionEditable: false,
+          preventDefaultEvents: false
         });
       }
     };
   }
-  componentDidMount() {}
 
   setScheduledDate(d) {
     this.props.setScheduledDate(this.props.id, d);
@@ -200,7 +260,10 @@ class Item extends Component {
       case "a":
         this.hotkeyHandler.ARCHIVE(event);
         return;
-      case "escape":
+      case "Enter":
+        this.hotkeyHandler.ENTER(event);
+        return;
+      case "Escape":
         this.hotkeyHandler.ESCAPE(event);
         return;
       case "r":
@@ -211,6 +274,9 @@ class Item extends Component {
         return;
       case "z":
         this.hotkeyHandler.UNDELETE(event);
+        return;
+      case "e":
+        this.hotkeyHandler.EDIT(event);
         return;
     }
   }
@@ -225,13 +291,17 @@ class Item extends Component {
           id={this.props.id}
           tabIndex="0"
           archived={this.props.archived}
+          itemType={this.props.type}
         >
           <ItemType itemType={this.props.type}>
             {this.props.completed ? "DONE" : this.props.type}
           </ItemType>
-          <ItemBody completed={this.props.completed}>
-            {this.props.type == "NOTE" &&
-              format(this.props.createdAt, "dd/MM/yyyy") + " - "}
+          <ItemBody
+            ref={c => (this.description = c)}
+            contentEditable={this.state.descriptionEditable}
+            completed={this.props.completed}
+            suppressContentEditableWarning={true}
+          >
             {removeItemTypeFromString(this.props.text)}
           </ItemBody>
           {this.props.dueDate && (
@@ -275,6 +345,9 @@ class Item extends Component {
 
 const mapStateToProps = state => ({});
 const mapDispatchToProps = dispatch => ({
+  updateItemDescription: (id, itemId) => {
+    dispatch(updateItemDescription(id, itemId));
+  },
   uncompleteItem: id => {
     dispatch(uncompleteItem(id));
   },
